[WEEK3][5] 100%

+++ Main/src/week3/Q5GasStation.java
@@ -7,8 +7,8 @@ public class Q5GasStation {
 
 	private static int	N;
 	
-	private static int[]	road;
-	private static int[]	gas;
+	private static long[]	road;
+	private static long[]	gas;
 	
 	public static void main(String[] args) throws IOException {
 		
@@ -24,13 +24,13 @@ public class Q5GasStation {
 		
 		N = Integer.parseInt(br.readLine());
 		
-		road = new int[N-1];
+		road = new long[N-1];
 		st = new StringTokenizer(br.readLine());
 		for (int i = 0; i < N-1; i++) {
 			road[i] = Integer.parseInt(st.nextToken()); 
 		}
 		
-		gas = new int[N-1];
+		gas = new long[N-1];
 		st = new StringTokenizer(br.readLine());
 		for (int i = 0; i < N-1; i++) {
 			gas[i] = Integer.parseInt(st.nextToken());
@@ -40,9 +40,9 @@ public class Q5GasStation {
 	
 	private static void sol() {
 
-		int cur;
-		int lowest = gas[0];
-		int total = gas[0] * road[0];
+		long	cur;
+		long 	lowest = gas[0];
+		long	total = gas[0] * road[0];
 		
 		for (int i = 1; i < N-1; i++) {
 			cur = gas[i];
